Moved chart package as per PR comments
