+++ api/src/main/java/org/pmiops/workbench/cohortbuilder/chart/ChartService.java
@@ -1,7 +1,6 @@
-package org.pmiops.workbench.chart;
+package org.pmiops.workbench.cohortbuilder.chart;
 
 import java.util.List;
-import org.pmiops.workbench.db.model.DbCohort;
 import org.pmiops.workbench.model.AgeType;
 import org.pmiops.workbench.model.CohortChartData;
 import org.pmiops.workbench.model.DemoChartInfo;
@@ -14,8 +13,6 @@ import org.pmiops.workbench.model.SearchRequest;
 public interface ChartService {
 
   List<CohortChartData> findCohortChartData(SearchRequest searchRequest, Domain domain, int limit);
-
-  List<CohortChartData> findCohortChartData(DbCohort dbCohort, Domain domain, int limit);
 
   List<DemoChartInfo> findDemoChartInfo(
       GenderOrSexType genderOrSexType, AgeType ageType, SearchRequest request);
