+++ api/src/test/java/org/pmiops/workbench/api/CohortBuilderControllerTest.java
@@ -36,12 +36,12 @@ import org.pmiops.workbench.cdr.model.DbCriteriaMenu;
 import org.pmiops.workbench.cdr.model.DbDomainCard;
 import org.pmiops.workbench.cdr.model.DbPerson;
 import org.pmiops.workbench.cdr.model.DbSurveyModule;
-import org.pmiops.workbench.chart.ChartQueryBuilder;
-import org.pmiops.workbench.chart.ChartService;
-import org.pmiops.workbench.chart.ChartServiceImpl;
 import org.pmiops.workbench.cohortbuilder.CohortBuilderService;
 import org.pmiops.workbench.cohortbuilder.CohortBuilderServiceImpl;
 import org.pmiops.workbench.cohortbuilder.CohortQueryBuilder;
+import org.pmiops.workbench.cohortbuilder.chart.ChartQueryBuilder;
+import org.pmiops.workbench.cohortbuilder.chart.ChartService;
+import org.pmiops.workbench.cohortbuilder.chart.ChartServiceImpl;
 import org.pmiops.workbench.cohortbuilder.mapper.CohortBuilderMapper;
 import org.pmiops.workbench.cohortbuilder.mapper.CohortBuilderMapperImpl;
 import org.pmiops.workbench.config.WorkbenchConfig;
